pyproject.toml
[project]
name = "llm-ai-service"
version = "0.1.0"
description = "大语言模型项目"
authors = [
    {name = "Quill", email = "quillnk@qq.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi (>=0.117.1,<0.118.0)",
    "uvicorn (>=0.36.0,<0.37.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "cryptography (<45)",
    "bcrypt (<4.0.0)",
    "resend (>=2.14.0,<3.0.0)",
    "celery[redis] (==5.5.3)",
    "redis (==5.0.6)",
    "flower (>=2.0.1,<3.0.0)",
    "eventlet (>=0.40.3,<0.41.0)",
    "minio (>=7.2.16,<8.0.0)",
    "qdrant-client (>=1.15.1,<2.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-json-logger (>=4.0.0,<5.0.0)",
    "sentence-transformers (>=5.1.2,<6.0.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "prometheus-client (>=0.23.1,<0.24.0)",
    "pymupdf (<1.24.0)",
    "unstructured (<0.16.0)",
    "pdfplumber (>=0.11.8,<0.12.0)",
    "python-docx (>=1.2.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
